feat(elysium): #1 - merge player stats feature
* feat(elysium): add API endpoint for getting player data

The `/api/playerinfo/` endpoint gives information on the player passed through the parameter. It first gets data from the Player Analytics API, available on the server, and then returns it in a stripped-down format.

* feat(elysium): add player stats page and player search feature

The new `/player` page features a search bar where users can search for a player. This will take them to `/player/stats?player=input` where they can view the stats of the inputted player. This is made possible through the Player Analytics API.